Move FakeRegistry to Mocks dir; Change include paths; Delete MetricsCollector
