+++ tests/Mocks/FakeRegistry.h
