+++ tests/Replication2/ReplicatedState/FollowerWaitForTest.cpp
@@ -27,7 +27,7 @@
 
 #include "Logger/LogMacros.h"
 #include "LogLevels.h"
-#include "Metrics/FakeRegistry.h"
+#include "Mocks/FakeRegistry.h"
 
 #include "Replication2/Mocks/FakeFollower.h"
 #include "Replication2/Mocks/FakeReplicatedState.h"
